Refactor how deployment works
This boy is a chonker. I pulled out all the deployment logic into
a `deployment.ActionManager` object. It's a lot cleaner and now
handles a lot of cases we weren't handling before. (Like startup!)

It isn't perfect and the error handling is probably problematic,
so let's call this WIP. I also fixed a lot of little nits that
were bothering me during the refactor so the diff is kinda
unnessesarily large.