Refactor how deployment works (#29)
* Refactor how deployment works

This boy is a chonker. I pulled out all the deployment logic into
a `deployment.ActionManager` object. It's a lot cleaner and now
handles a lot of cases we weren't handling before. (Like startup!)

It isn't perfect and the error handling is probably problematic,
so let's call this WIP. I also fixed a lot of little nits that
were bothering me during the refactor so the diff is kinda
unnessesarily large.

* Fix linter problems with ActionManager

* Remove `godox`, which is not enabled on CI

* Remove `wsl`, which CI also doesn't know about

* API Handler for changing component state (#30)

* API Handler for changing component state

* Linting nit fix

* Name changes and linter fix

* Swapy

* Fix deploying from wrong branch, rename api handler

* Use pigz, create new components in the 'not_a_component' state

* Add database pushing of what components are currently deployed

* Finish the removal of unnesesary pointers to IDs

* Fix comment typo

Co-authored-by: Henry Jaensch <hjaensch7@gmail.com>